:recycle: Split `WrapErrorFuture` to mod `handler::future`
